// setup/arcgis-experience-builder-1.6/ArcGISExperienceBuilder/client/dist/widgets/common/search/src/config.ts
export enum SearchServiceType {
  GeocodeService = 'GeocodeService',
  FeatureService = 'FeatureService'
}
// ...
export interface LayersConfig {
  configId: string
  label: string
  useDataSource: UseDataSource
  icon?: any
  searchFields?: FieldSchema[]
  searchExact?: boolean
  displayFields?: FieldSchema[]
  hint?: string
  searchServiceType?: SearchServiceType
}

export interface GeocodeConfig {
  configId: string
  label: string
  geocodeURL: string
  outputDataSourceId: string
  useDataSource: UseDataSource
  icon?: any
  hint?: string
  searchServiceType?: SearchServiceType
}

export type IMGeocodeConfig = ImmutableObject<GeocodeConfig>

export type DatasourceConfigItem = LayersConfig & GeocodeConfig